CHG: ReportViewerHandler returns 404 if the file system report is unable to locate requested file.

+++ Origam.Server/ReportViewerHandler.cs
@@ -43,6 +43,7 @@ along with ORIGAM.  If not, see<http://www.gnu.org/licenses/>.
 using System;
 using System.Collections;
 using System.IO;
+using System.Net;
 using System.Web;
 using System.Web.SessionState;
 using log4net;
@@ -141,11 +142,18 @@ namespace Origam.Server
             string filePath = ReportHelper.BuildFileSystemReportFilePath(
                 report.ReportPath, reportRequest.Parameters);
             string mimeType = HttpTools.GetMimeType(filePath);
+            if (!File.Exists(filePath))
+            {
+                context.Response.StatusCode = (int)HttpStatusCode.NotFound;
+                context.Response.End();
+                return;
+            }
             context.Response.ContentType = mimeType;
             context.Response.AddHeader(
                 "content-disposition",
-                /*"attachment; " +*/ httpTools.GetFileDisposition(
-                    new FxHttpRequestWrapper(context.Request), Path.GetFileName(filePath)));
+                httpTools.GetFileDisposition(
+                    new FxHttpRequestWrapper(context.Request), 
+                    Path.GetFileName(filePath)));
             context.Response.WriteFile(filePath);
         }
         private void HandleWebReport(
